+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2026-07-13
+
+### Added
+
+- Added support for project-wide configuration overrides inside `package.json` under the `"envrepair"` block (allowing custom target and template files, and default environment modes).
+- Added an integration test checking configuration overrides in `package.json`.
+
 ## [0.1.9] - 2026-07-13
 
 ### Added